requirements.txt
python-telegram-bot==20.7
feedparser==6.0.11
requests==2.31.0
beautifulsoup4==4.12.3
apscheduler==3.10.4
python-dotenv==1.0.0

